[#1087] Renamed and moved a few CommitPanel components.
