+++ frontend/src/commit-panel/CommitPanel.less
