+++ frontend/src/commit-panel/Notice.tsx
