+++ frontend/src/commit-panel/details/Details.tsx
